refactor: move statistics component to separate file
