+++ src/view/show-more.js
@@ -1,0 +1,7 @@
+export {renderShowMoreCompnent};
+
+const renderShowMoreCompnent = () => {
+  return `
+    <button class="films-list__show-more">Show more</button>
+  `;
+};
