+++ src/view/statistics.js
@@ -1,0 +1,9 @@
+export {renderStatisticsComponent};
+
+const renderStatisticsComponent = () => {
+  return `
+    <section class="footer__statistics">
+      <p>130 291 movies inside</p>
+    </section>
+  `;
+};
